Reorganization of settings files
