+++ lib/define_env.dart
@@ -3,6 +3,10 @@ library define_env;
 export 'src/android_studio_config_writer.dart';
 export 'src/clipboard.dart';
 export 'src/define_env.dart';
-export 'src/define_env_settings.dart';
-export 'src/env_settings.dart';
+export 'src/settings/check_env.dart';
+export 'src/settings/check_env_settings.dart';
+export 'src/settings/entities/env_settings.dart';
+export 'src/settings/entities/field_env_settings.dart';
+export 'src/settings/entities/field_type_env_settings.dart';
+export 'src/settings/load_env_settings.dart';
 export 'src/vscode_config_writer.dart';
