+++ lib/src/settings/entities/field_type_env_settings.dart
@@ -1,0 +1,10 @@
+import 'package:json_annotation/json_annotation.dart';
+
+enum FieldTypeEnvSettings {
+  @JsonValue('String')
+  string,
+  bool,
+  int,
+  @JsonValue('enum')
+  enum$,
+}
